Models

Description

Models in our db, and possibly in the zabbix db. Also provides migrateAll that migrates an empty or partial db schema to ours.

__Note that we do not provide any indices, that are crucial for multi-axis-graphs dashboard performance.__ Refer to that documentation no how to set them up.

Documentation

data Group

Constructors

Group

Fields

groupName :: !Text

type GroupId = Key Group

data Host

Constructors

Host

Fields

hostHost :: !Text

hostStatus :: !Int

hostAvailable :: !Int

hostName :: !Text

type HostId = Key Host

data HostGroup

Constructors

HostGroup

Fields

hostGroupHost :: !(Key Host)

hostGroupGroup :: !(Key Group)

type HostGroupId = Key HostGroup

data Application

Constructors

Application

Fields

applicationHost :: !(Key Host)

applicationName :: !Text

type ApplicationId = Key Application

data Item

Constructors

Item

Fields

itemType :: !Int

itemHost :: !(Key Host)

itemName :: !Text

itemKey_ :: !Text

itemDescription :: !Text

itemValueType :: !Int

type ItemId = Key Item

data ItemApp

Constructors

ItemApp

Fields

itemAppApp :: !(Key Application)

itemAppItem :: !(Key Item)

type ItemAppId = Key ItemApp

data History

Constructors

History

Fields

historyItem :: !(Key Item)

historyClock :: !Int

historyValue :: !Rational

type HistoryId = Key History

data Trend

Constructors

Trend

Fields

trendItem :: !(Key Item)

trendClock :: !Int

trendValueMin :: !Rational

trendValueAvg :: !Rational

trendValueMax :: !Rational

type TrendId = Key Trend

data FutureModel

Constructors

FutureModel

Fields

futureModelName :: !Text

type FutureModelId = Key FutureModel

data ItemFuture

Constructors

ItemFuture

Fields

itemFutureItem :: !(Key Item)

itemFutureModel :: !(Key FutureModel)

itemFutureParams :: !ByteString

itemFutureDetails :: !ByteString

itemFutureIsMaster :: !Bool

type ItemFutureId = Key ItemFuture

data Future

Constructors

Future

Fields

futureItem :: !(Key ItemFuture)

futureClock :: !Int

futureValue :: !Rational

type FutureId = Key Future

data Threshold

Constructors

Threshold

Fields

thresholdItem :: !(Key ItemFuture)

thresholdLower :: !Bool

thresholdHigh :: !Double

thresholdWarning :: !Double

thresholdCritical :: !Double

type ThresholdId = Key Threshold

data Metric

Constructors

Metric

Fields

metricName :: !Text

metricKey_ :: !Text

metricScale :: !Int

type MetricId = Key Metric

migrateAll :: Migration